move random functions to random.c
