+++ ext/mri/cmodules/random.c
@@ -1,0 +1,45 @@
+static VALUE arf_create_random_engine(VALUE self){
+  return Qnil;
+}
+static VALUE arf_retain_random_engine(VALUE self){
+  return Qnil;
+}
+static VALUE arf_random_engine_set_type(VALUE self){
+  return Qnil;
+}
+static VALUE arf_random_engine_get_type(VALUE self){
+  return Qnil;
+}
+static VALUE arf_random_uniform(VALUE self){
+  return Qnil;
+}
+static VALUE arf_random_normal(VALUE self){
+  return Qnil;
+}
+static VALUE arf_random_engine_set_seed(VALUE self){
+  return Qnil;
+}
+static VALUE arf_get_default_random_engine(VALUE self){
+  return Qnil;
+}
+static VALUE arf_set_default_random_engine_type(VALUE self){
+  return Qnil;
+}
+static VALUE arf_random_engine_get_seed(VALUE self){
+  return Qnil;
+}
+static VALUE arf_release_random_engine(VALUE self){
+  return Qnil;
+}
+static VALUE arf_randu(VALUE self){
+  return Qnil;
+}
+static VALUE arf_randn(VALUE self){
+  return Qnil;
+}
+static VALUE arf_set_seed(VALUE self){
+  return Qnil;
+}
+static VALUE arf_get_seed(VALUE self){
+  return Qnil;
+}
